UPSTREAM: amd/fam10: Add k10temp ACPI thermal zone mixin.
This ACPI thermal zone is applicable to AMD family 10 to 14 (and some
15) CPUs.

It should not be used on boards for which errata 319 (The thermal sensor
of Socket F/AM2+ processors may be unreliable) is applicable. AM3 and
later should be fine.

Derived from src/northbridge/amd/amdk8/thermal_mixin.asl

Change-Id: I0d69c558f5aea526fb818259ab04a93b2ff78598
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: d5e6618a4f076610e683b174c4dd5108d960c785
Original-Change-Id: Id036cbf4cd717c3320a720edc452945df2b5e072
Original-Signed-off-by: <ranma+coreboot@tdiedrich.de>
Original-Reviewed-on: http://review.coreboot.org/10617
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Timothy Pearson <tpearson@raptorengineeringinc.com>
Reviewed-on: https://chromium-review.googlesource.com/288050
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Commit-Queue: Stefan Reinauer <reinauer@chromium.org>
Tested-by: Stefan Reinauer <reinauer@chromium.org>