UPSTREAM: azalia: fix up and clean up shrinkage of boilerplate code
Should fix regression in HDA verb setup on nvidia mcp55 and intel sch
southbridges.  The mcp55 code could not find the mainboard's verb table
because the table was not even being compiled in.  The sch boards appeared
to have the same issue.

Intel broadwell and fsp_bd82x6x seemed to have not gotten the boilerplate
shrink, so apply it to those too.

Followup-to: Ib3e09644c0ee71aacb067adaa85653d151b52078
             (azalia: Shrink boilerplate)

Change-Id: I2eff648cbd0cd24ad4380892530e090b151bc2de
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: ec505ad21c923c114a16b2710a0113f657765430
Original-Change-Id: If7aae69f5171db67055ffe220bdff392caaa5d9f
Original-Signed-off-by: Jonathan A. Kollasch <jakllsch@kollasch.net>
Original-Reviewed-on: http://review.coreboot.org/10826
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/288055
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Commit-Queue: Stefan Reinauer <reinauer@chromium.org>
Tested-by: Stefan Reinauer <reinauer@chromium.org>