UPSTREAM: libpayload: lz4: Add output overrun check to incompressible case
The LZ4 decompressor currently doesn't check for output overruns before
writing data in the case where a block had been incompressible (and
included verbatim in the compression stream). This is extremely unlikely
with the default 4MB blocks, but still a nice thing to fix. We'll still
output as much data as we can before returning an error to support
partial decompression use cases.

This matches the behavior already in place for normal, LZ4-compressed
blocks where the decompression function is already (supposed to be)
doing complete bounds checking (although it is not guaranteed to output
all valid bytes before aborting on an output overrun, and you should try
to provide a few dozen bytes of extra buffer space beyond the parts
you're interested in on partial decompression).

BRANCH=None
BUG=chrome-os-partner:32184
TEST=None

Change-Id: If132b729a2223bbafad23107ea9b9b8011aa2631
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: bf697566da9a59a37d1bf3dfa1839fd5aee33f56
Original-Change-Id: I5e40c8cec8947ec0ec8f6d8c8fa2574cfb4dc958
Original-Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Original-Commit-Id: 636985334c9b3b93a12d4066d2829f1f999c9315
Original-Original-Change-Id: Iecf44650aade60b9fa1b13e57da752fb482a3f3f
Original-Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Original-Reviewed-on: https://chromium-review.googlesource.com/286240
Original-Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-on: http://review.coreboot.org/11016
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-on: https://chromium-review.googlesource.com/288138