glados: make EC_SMI_L functional
Set the EC_SMI_GPI define to be GPP_E15 and route that
GPIO for SMI generation. Also, the mainboard_smi_gpi_handler()
was introduced on skylake in order to process any GPI that could
generate an SMI. Switch to this handler so one can process the
appropriate events.

BUG=chrome-os-partner:43778
BRANCH=None
TEST=Used 'lidclose' on EC command line during depthcharge
     to confirm EC_SMI_L generates SMI and shutdown happens.

Change-Id: Ia365b86161670a809e3fa99dde38fccc612d5e77
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/291934
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>