UPSTREAM: glados: make EC_SCI_L work
This actually just incorporates a comment change.

Original description:

In order for the EC_SCI_L to work the GPE0 route needs
to be set along w/ the GPE event for the EC. As the GPE0
route is dynamic the EC_SCI_GPI needs to be set along
with the route so everything lines up. In this case, the
GPE0 route is set to the defaults such that GPP_C, GPP_D,
and GPP_E are routed to GPE0 block 0, 1, and 2, respectively.
This works out for glados because the EC_SCI_L is connected
to GPP_E16.

Original-BUG=chrome-os-partner:43778
Original-BRANCH=None
Original-TEST=Built and booted glados. The 'acpi' interrupt in /proc/interrupts
     is incrementing as well as /sys/firmware/acpi/interrupts/gpe50.

Original-Change-Id: I71fc4bec124f3ac87453a099412154e67aba6280
Original-Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/292011
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>

Original-Change-Id: Idbb6d29364655537abc9ae6f012b3abb38edf138
Original-Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-on: http://review.coreboot.org/11210
Original-Tested-by: build bot (Jenkins)

BUG=None
BRANCH=None
TEST=Built glados.

Change-Id: Ib1696b1d204503a1507890a08b8ca069885ef681
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/294460
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>