veyron_rialto: Turn on all leds
Without this, the leds would be stuck to whatever the pullup/down states the
pins come with on rk3288.

Ready2_LED, an orange led, is one of the leds in this state.
This might confuse some users thinking there's an error.

Turn all of them on instead.
Later on depthcharge will use the same LEDs to indicate dev mode status.

BUG=chrome-os-partner:44274
BRANCH=master
TEST=Boot firmware without anything else, note all leds on

Change-Id: I4c4e8940dd9cf1ac0301ac00bfc5992ba16e1589
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/294065
Reviewed-by: David Hendricks <dhendrix@chromium.org>