glados: fix kepler probing
On proto2 boards the kepler device has its reset line pulled up
to one of its IO rails with a zener in between. This results in the
device not being visible at MemoryInit() time because for some
reason FSP is doing PCIE configuration/probing in that path. Hack
around the broken FSP logic by configuring the pads for kepler's
power and clkreq.

BUG=chrome-os-partner:44326
BRANCH=None
TEST=Built and booted glados. lscpi shows the device on bus 2.

Change-Id: I7fe4a707f9321b7bdec4b4be729c5d0dcce65f6e
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/294810
Reviewed-by: Robbie Zhang <robbie.zhang@intel.com>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>