UPSTREAM: cbfstool: pass cbfs_file header into "compress" functions
These functions can do all kinds of things, such as converting an ELF
image into SELF, or (in the future) compress or checksum entire files.
This may require changing or adding fields to the header, so they
need to have access to it.

The header_size parameter that was provided (but never used) is
equivalent to cbfs_file's offset field.

Change-Id: Ic068000661f1464227a9ed198caad0001c172e89
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: 056f6a1f80952d91e0aa41d379ae9f08032595cf
Original-Change-Id: I7c10ab15f3dff4412461103e9763a1d78b7be7bb
Original-Signed-off-by: Patrick Georgi <patrick@georgi-clan.de>
Original-Reviewed-on: http://review.coreboot.org/11325
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/301637
Commit-Ready: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>