UPSTREAM: linking: lay the groundwork for a unified linking approach
Though coreboot started as x86 only, the current approach to x86
linking is out of the norm with respect to other architectures.
To start alleviating that the way ramstage is linked is partially
unified. A new file, program.ld, was added to provide a common way
to link stages by deferring to per-stage architectural overrides.
The previous ramstage.ld is no longer required.

Note that this change doesn't handle RELOCATABLE_RAMSTAGE
because that is handled by rmodule.ld. Future convergence
can be achieved, but for the time being that's being left out.

BUG=chrome-os-partner:44827
BRANCH=None
TEST=Built a myriad of boards.

Original-Change-Id: I5d689bfa7e0e9aff3a148178515ef241b5f70661
Original-Signed-off-by: Aaron Durbin <adubin@chromium.org>
Original-Reviewed-on: http://review.coreboot.org/11507
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Tested-by: Raptor Engineering Automated Test Stand <noreply@raptorengineeringinc.com>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>

Change-Id: If0101ff8eb31229bf25000855bc63e9fd78c00e6
Signed-off-by: Aaron Durbin <adubin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/298723
Commit-Ready: Aaron Durbin <adurbin@chromium.org>
Tested-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>