UPSTREAM: rules.h: add fall through where no ENV_<STAGE> is set
There are cases where rules.h can be pulled in, but the
usage is not associated with a particular stage. For
example, the cpu/ti/am335x build creates an opmap header.
That is a case where there is no stage associated with
the process. Therefore, provide a case of no ENV_>STAGE>
being set.

BUG=chrome-os-partner:44827
BRANCH=None
TEST=Built a myriad of boards. Analyzed readelf output.

Original-Change-Id: Ia9688886d445c961f4a448fc7bfcb28f691609db
Original-Signed-off-by: Aaron Durbin <adubin@chromium.org>
Original-Reviewed-on: http://review.coreboot.org/11513
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>

Change-Id: I6ddc78dbad21e24ebedd7c4be02a8a1b65559519
Signed-off-by: Aaron Durbin <adubin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/298721
Commit-Ready: Aaron Durbin <adurbin@chromium.org>
Tested-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>