UPSTREAM: microcode: Unify rules to add microcode to CBFS once again
Now that cbfstool supports file alignment, we can use the conveniently
available <filename>-align handler, and remove the need to have a
separate rule in src/Makefile.inc just for adding the microcode.

We can also get rid of the layering violation of having the
CONFIG_PLATFORM_USES_FSP1_0 symbol in a generic src/cpu/ makefile.
Note that we still have a layering violation by the use of the
CONFIG_CPU_MICROCODE_CBFS_LOC symbol, but this one is acceptable
for the time being.

BUG=None
BRANCH=None
TEST=Built and booted glados.

Original-Change-Id: Id2f8c15d250a0c75300d0a870284cac0c68a311b
Original-Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>
Original-Reviewed-on: http://review.coreboot.org/11526
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>

Change-Id: I9a6b8765272e39d6aebb0130e3a32a4e7decf7d3
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/298716
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>