UPSTREAM: intel: Do not hardcode the position of mrc.cache
The reason for hardcoding the position of the MRC cache was to satisfy
the alignment to the erase size of the flash chip. Hardcoding is no
longer needed, as we can specify alignment directly. In the long term,
the MRC cache will have to move to FMAP, but for now, we reduce
fragmentation in CBFS.

Note that soc/intel/common hardcoding of mrc.cache is not removed, as
the mrc cache implementation there does not use CBFS to find the cache
region, and needs a hardcoded address.

BUG=None
BRANCH=None
TEST=Built and booted glados.

Original-Change-Id: I5b9fc1ba58bb484c7b5f687368172d9ebe625bfd
Original-Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>
Original-Reviewed-on: http://review.coreboot.org/11527
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>

Change-Id: Icba82366aece4179fd8622ed6686f404ce1dd266
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/298715
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>