UPSTREAM: x86: provide minimum alignment for romstage
The current way the XIP address of romstage is calculated is by
doing a 'cbfstool locate' using a bin file of romstage linked
at address 0. That address is then used for re-linking romstage at
the address spit out by cbfstool. Currently, the linker actually
sets minimum alignment on the text sections as 32 bytes, but it
doesn't actually honor that value. Instead, provide a minimum
alignment for romstage so as not to fight the linker.

BUG=chrome-os-partner:44827
BRANCH=None
TEST=Built asus/kfsn4-dre. Confirmed ROMSTAGE_BASE == gdtptr.

Original-Change-Id: Id6ec65d257df9ede78c720b0d7d4b56acfbb3f15
Original-Signed-off-by: Aaron Durbin <adubin@chromium.org>
Original-Reviewed-on: http://review.coreboot.org/11588
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>

Change-Id: I696746d39ae1378337ff7f92f229f3fdf0d88c7f
Signed-off-by: Aaron Durbin <adubin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/298722
Commit-Ready: Aaron Durbin <adurbin@chromium.org>
Tested-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>