skylake/braswell: protect against struct packing policy
Intel's code in FSP and EDK uses #prgram pack(1) without
pushing/popping the current state. The result is that 2
different compilation units can have different views of
an unaffiliated  struct's layout depending on the ordering
of Intel's headers. In order to protect against this scenario
the current packing policy is pushed and popped around the
Intel headers.

BUG=chrome-os-partner:43522
BRANCH=None
TEST=Built and booted glados.

Change-Id: I5380b83461de4db09a24287214b1e2f9892db7d8
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/298480
Reviewed-by: AlexandruX Gagniuc <alexandrux.gagniuc@intel.com>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>