UPSTREAM: cbfstool: remove locate command
The locate command was previously being used for x86 romstage
linking as well as alignment handling of files. The add command
already supports alignment so there's no more users of the
locate command. Remove the command as well as the '-T' (top-aligned)
option.

Note: This patch uses a 0 metadata size in the 'cbfstool add' path
      because the compression metadata patches are not in chromium.

BUG=chrome-os-partner:44827
BRANCH=None
TEST=Built rambi. Noted microcode being directly added.

Original-Change-Id: I3b6647bd4cac04a113ab3592f345281fbcd681af
Original-Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-on: http://review.coreboot.org/11671
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>

Change-Id: I67e023d014331cf980825f1de32582e40298e5d5
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/300056
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>