UPSTREAM: cbfstool: make fmap search more strict
Since fmap doesn't come with a checksum, we resort to a number of
heuristics to determine if a given location hosts an fmap (instead of
another data structure that happens to store the fmap magic string at
the right location).

The version test is particularly effective against strings containing
the magic (which either terminate with 0, or have some other ASCII data,
but rarely a '\001' byte inside the string).

Change-Id: Ie35af143f81d04929d1187da065312973b79f30f
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: b946f12ed0c4a4177b64c032f17dd5b3fd2a0c99
Original-Change-Id: Ic66eb0015c7ffdfe25e0054b7838445b8ba098e9
Original-Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Reviewed-on: http://review.coreboot.org/11690
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/301660
Commit-Ready: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>